refactor(common): extrai infraestrutura Spring para módulo common-spring
- Move FeignHeadersConfig e HTTP filters/interceptors para common-spring (opcional)
- Mantém common apenas com DTOs/eventos e ApiError (sem dependência de Spring)
- Atualiza root POM e adiciona novo módulo common-spring
- Ajusta POMs de api-gateway, task-service e activity-service para depender de common-spring
- Garante build bem sucedido em multi-módulo
