+++ common/src/main/java/com/viniss/todo/common/http/RequestLoggingConfig.java
@@ -1,18 +1,1 @@
-package com.viniss.todo.common.http;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class RequestLoggingConfig implements WebMvcConfigurer {
-
-  @Autowired
-  RequestLoggingInterceptor interceptor;
-
-  @Override
-  public void addInterceptors(InterceptorRegistry registry) {
-    registry.addInterceptor(interceptor).addPathPatterns("/**");
-  }
-}
+// moved to common-spring
