+++ pom.xml
@@ -9,6 +9,7 @@
 
   <modules>
     <module>common</module>
+    <module>common-spring</module>
     <module>api-gateway</module>
     <module>task-service</module>
     <module>activity-service</module>
