+++ task-service/pom.xml
@@ -32,6 +32,11 @@
       <artifactId>common</artifactId>
       <version>0.1.0</version>
     </dependency>
+    <dependency>
+      <groupId>com.viniss.todo</groupId>
+      <artifactId>common-spring</artifactId>
+      <version>0.1.0</version>
+    </dependency>
       <dependency>
           <groupId>org.springframework.boot</groupId>
           <artifactId>spring-boot-starter-actuator</artifactId>
